UIforETW now measures save and merge times and warns if they're long.
There are various things that can cause the save or merge steps to go
long. Investigating these is always painful. The first step is to
measure them and, if Show Commands is checked, display the times.

The next step is to warn when they are long and suggest posting to
the UIforETW group.

The final step is to have a batch file, metatrace.bat, which can be
used to record a trace of a trace being recorded.

This means that if a user hits a problem they can turn on Show Commands,
share the output, then record a metatrace, and share that. Suddenly
these problems will become *easy* to investigate.