Improve warning messages if the merge step runs slow.
I tried to move the amcache.hve renaming step to a tighter scope
but it actually caused problems because apparently both the -flush
and the -merge steps are made slow by amcache.hve. Sigh...

I could measure the time for the -flush step and warn on that, but
I don't feel so inclined.