Clear TracingFlags when the heap tracing exe name is changed.
It is important to not leave TracingFlags==1 when heap tracing is
not supposed to be enabled, as it may lead to unwanted heap tracing
or extra overhead. This change makes sure that old heap settings
don't get orphaned and do get cleared.

This bug was found while writing the documentation for UIforETW
heap tracing.