Add support for loading ETW traces in WPA 10.0, turn off /analyze.
The build times with /analyze are too long. They were discouraging
me from making changes. Sorry. I will continue to occasionally run
code analysis to look for bugs.

Early versions of WPA 10.0 are available. I tried them out and
reported a few bugs, and found a couple of areas where the 8.1
startup profile disagrees with them:

1) The regex used in the Generic Events for displaying Multi-* and
Chrome* events is not supported in 10.0. This is a serious problem.
For now the startup10.wpaProfile hacks around this by defaulting to
displaying just Multi-* events.

2) "NewProcess" was renamed to "New Process" in the CPU Usage (Precise)
graph. Grrr. Churn. This triggers a warning and makes the graph not
show up.

So, I force load startup10.wpaProfile. However this doesn't work
brilliantly because the old startup profile also loads! Better than
nothing, but not great.