Tweak the ETWProvider project - saner output and add etwprof.h.
ETWProvider was using a post-build step to copy the .lib files to
the lib directory. This was obscure (it took me a while to find how
I had implemented this) and unreliable (switching between build
configurations and building won't necessarily copy the files) and
it wasn't copying the .dll files. All fixed.

And, I added etwprof.h to the project file, to make it easier to find.