Many working set tweaks and changes for better performance and compatibility.
The working set events now take wchar_t* for the process name, and unsigned
integer KiB counts instead of float MiB counts. The wchar_t* change allows
using CreateToolhelp32Snapshot, which is faster than EnumProcesses followed
by querying the modules for the EXE name. The changed to unsigned KiB is
for compatibility with the columns in Process Explorer.