Add a 64-bit build of UIforETW, necessary for 64-bit working sets.
The QueryWorkingSet function returns truncated results for some 64-bit
processes. My best guess is that it returns results for those pages that
are below the 4 GB limit, which is why some processes show up as having
tiny working sets and others are quite large.

This problem is undocumented except for some frustrated community additions
on the QueryWorkingSet page.