add reserve, swap push_back with emplace_back
In places where we'd fill a `std::vector` inside a for loop, I added a
call to `reserve` to preallocate the necessary space.

I replaced `push_back` with `emplace_back`, saving silly copy
construction.

In ranged-for loops, I changed `auto` to `auto&` to avoid copies, and
added const wherever the items are not modified.