Fix StripChromeSymbols.py to handle content.dll and blink_web.dll.
When building Chrome with "GYP_DEFINES=component=shared_library" the
set of DLLs created is much larger and very different from a shipping
Chrome build. This means that there are two additional DLLs that are
large enough to need stripping with StripChromeSymbols.py to avoid
excessive symbol translation times in WPA. This change is a quick
hack to add content.dll and blink_web.dll to the list. It also
improves the logic for renaming DLLs to/from, and uses a try/except
handler to make it more likely that renaming back will actually occur.

The code could use some cleanup, such as extracting the DLL name
selected from the output line instead of manually setting dllMatch.
Oh well. I just need it to work for now.