Fix IdentifyChromeProcesses.py to group browsers by type.
When working on Chrome I may end up with two or three versions
running simultaneously - the stable version that I use for email and
bug triage, and the locally built version that I am testing. Having
all of the PIDs grouped together (two browsers, two gpu-process) is
confusing, so now I group them by exePath.

This doesn't work perfectly if the test browser is launched through a
relative path, but that is a minor (and indeed avoidable) glitch.