Only look for chrome.exe in the exePath portion of the commandline.
IdentifyChromeProcesses.py was producing inaccurate results when
run on a trace that included building chrome. The text chrome.exe
was showing up in various build commands which were then being
interpreted as *being* chrome.exe. Now the logic is much tighter
and it only looks within the already parsed-out exePath.