Add power, frequency, and temperature monitoring with Intel Power Gadget.
Based on sample code at:
https://software.intel.com/en-us/blogs/2012/12/13/using-the-intel-power-gadget-api-on-mac-os-x
I got the power status class to look for and use Intel Power Gadget if it
is installed. This allows emittin events for the CPU power draw, temperature,
and frequency. The temperature and frequency are sampled at that point but the
power draw represents power consumption since the last sample and is, apparently, quite
accurate.

The documentation says that the temperature should be in Celsius but on some machines
I get results that are clearly Fahrenheit.

This should help with investigations of thermal throttling.