Disable the key logger thread when debugging to avoid hangs.
When debugging UIforETW there have always been hangs - whenever
UIforETW is paused in the debugger the entire system would take
about ten seconds to respond to input. This makes single-stepping
painful and alt+tabbing impossible.

I finally realized that this was caused by the input hooks. When
UIforETW is paused in the debugger the input hook thread can't
respond in a timely manner and the system has to timeout on each
input, about ten seconds at a time.

I now disable the input hook thread when a debugger is present at
startup.