Automatically use more trace buffers on machines that can afford it.
When tracing to a file it makes sense to have as many buffers as
possible to make sure that events do not get lost. With this change
the default buffer count is increased by 50% to 100% on machines with
15+ GB and 30+ GB respectively, using the new BufferCountBoost
function.

Adding this functionality necessitated stringPrintf which is a
generally handy function anyway. While adding that I added debugPrint
for printing to outputDebugString.

Tracing to memory is not affected - the buffer count there should
be user controlled to allow controlling the size of the resultant
traces.