Handle pasting of LF terminated text into the trace information
Windows edit controls insist on having CR-LF terminated. If you
have LF terminated text in the clipboard then it gets pasted as one
long line. Since Chromium's standard is to have LF terminated
source code this can definitely be a problem.

Fixing this required plumbing through a few things:

A ConvertToCRLF function needed to be written.
An accelerator and handler for Ctrl+V needed to be hooked up to the
trace information edit control.
A GetClipboardText function was needed.

But, hey, useful functions to have. And now vprintf can use
ConvertToCRLF instead of its previous hacky and buggy loop.