Pass stdInput to subprocesses and start using Python's subprocess.
On some machines (Python 3 installs?) and when using the subprocess
module Python scripts launched by UIforETW would fail with:
    "WindowsError: [Error 6] The handle is invalid"

That's because hStdInput was not being set before calling CreateProcess.
This change sets hStdInput and switches to the subprocess module.

It also sets stderr to subprocess.STDOUT since otherwise xperf's
stderr is missed by the Python script and then captured and displayed
by UIforETW, which is undesirable. Redirecting it to STDOUT causes it
to be swallowed by the Python script.