Auto-save/stop tracing to file if it runs for 'too long'.
If tracing to a file runs for more than 120 seconds then tracing is
automatically stopped. This is a good thing because long traces are
typically unwieldy and unusable, and if tracing is started and then
forgotten it can eventually fill your hard drive (I accidentally
recorded a 70 GB file in this manner).

The timeout should be configurable, but it isn't.