Finish controlling of Chrome's ETW event filtering from UIforETW
- Added a link to the code review that added ETW event filtering to Chrome.
- Added support for the special all enabled-by-default and all disabled-by-default
categories.
- Properly deciding when to register the Chrome providers and properly passing the
chromeKeywords_ to xperf.
- Minor tweaks to the settings dialog.

The Chrome settings should really all be moved to a separate dialog. Some other day.

Note that event filtering is only supported in *very* new versions of Chrome (the
Chrome side change was checked in June 25, 2015, see https://codereview.chromium.org/1176243016)
and it currently requres this command-line switch to Chrome: --trace-export-events-to-etw

It also requires that tracing be running when Chrome is launched. These last two
restrictions will be removed eventually through changes to Chrome.