Add monitoring of Windows timer interval to ETW traces.
The Windows timer interval can be altered by any program and this
can change the performance and power efficiency of the entire system.
Therefore it seems appropriate to monitor the current timer interval.

To do this a CPowerStatusMonitor::SampleTimerState function was added
which is called five times a second on the power status thread. It calls
the undocumented NtQueryTimerResolution to get the current timer
resolution, and calls the newly added ETWMarkTimerInterval function to
record this to the trace.