Improve support for WPA 10 now that it co-exists with WPA 8.1
WPA 10 saves its settings and startup profile to a different location
so it can now co-exist with WPA 8.1. With this change I:

- Updated the startup10.wpaProfile file to work correctly with the
latest (10.0.10158.0) version of WPA. This included fixing CPU Usage
(Precise) enablement to "New Process" instead of "NewProcess", and
resetting the GPU graph.
- Added code to copy both startup profiles to their destinations, both
at startup (if the startup profile is missing) and from the settings
dialog (forced by the user).
- Centralized the profile copying code in utility.cpp

Previous versions of WPA 10 should not be used because they break things
like the Chrome/Multi filtering in generic events, but I don't bother
checking for old versions.

WPA 8.1 is still the default and is in fact still required, but that will
change soon.