Update UIforETW to handle when WPT 10 is installed but not WPT 8.1
WPT 10 now works well enough that requiring WPT 8.1 is inappropriate.

UIforETW still defaults to using WPT 8.1 but if it does not exist
then it will point wptDir_ at the 10.0 version. The error messages
and popup menu enabling/disabling were updated.

This also fixes some dodgy code that left a dangling progFilesx86Dir
pointer after the memory was freed, and fixes a bug whereby the
"Open trace in WPA 10" menu was not disabled when no trace was
selected.