Fix StripChromeSymbols.py to handle PDB 'age' greater than 15
The xperf symcache action with the -dbg option prints the GUID and
age for each PDB referenced in an ETW trace file. This lets me
determine whether the corresponding .symcache file exists. However
the the symcache action prints the age in *decimal* and the .symcache
files contain it printed in *hexadecimal*. This means that
StripChromeSymbols.py will report that there are missing .symcache
files if you have PDBs whose age is greater than 15, even if their
.symcache files exist.