Command-line tool for monitoring Intel power MSRs
The CPowerStatusMonitor class is mostly a copy of the same class in
UIforETW, modify to suit the command-line environment. Generalizing it
to allow code sharing is left as an exercise for the reader, but doesn't
really seem worth it.

This tool prints various power and temperature stats occasionally. There
are no options.