Check the return code of SHGetKnownFolderPath
/analyze warns if you use results from a function without checking
whether that function succeeded. This can be good, especially for
functions like fread, but it is annoying for functions that should
never fail. It certainly makes me wish that Windows had more no-fail
contracts. If SHGetKnownFolderPath fails then that suggests Windows
is busted and I don't want to try to 'handle' that. So, I simply
check and then std::terminate.