Finalizing Windows 10 and WPT 10 support.
Windows 10 requires that WPT 10 be used to change the CPU sampling
speed. This is now enforced.

Windows 7 requires that WPT 8.1 be used for recording traces because
WPT 10 fails to record debug information that is needed for symbol
resolution. This is now enforced.

This change also fixes an output statement so that the time to save
a trace is not displayed when you are just stopping tracing.