Squashed commit of PowerStatus fixes from Alexander Riccio
/analyze complained:
C28182 Dereferencing a copy of a null pointer 'pdidd' contains the same

Fixed by using std::vector which also simplifies error handling
Created HDEVINFO_battery and and Battery objects to further manage
resources and simplify error handling.

There were so many levels of indentation in
`CPowerStatusMonitor::SampleBatteryStat` that it was giving me a
headache, so I went ahead and 'unrolled' it.

Because I'm crazy about checking return values, I added a wrapper
function to `Utility.h`/`Utility.cpp` (`CloseValidHandle`), as if the
wrapped system call ever fails, there's a bug on our side of things.

As a drive-by fix, I also changed the `OUT` pseudo-annotated parameters
of `NtQueryTimerResolution` (`OUT` is defined to `` - blank text) to
`_Out_`, so that `/analyze` treats them as actual out parameters.

There was a bogus C6102 that I (brucedawson) suppressed.