Set generic event aggregation modes to average.
In order to graph custom data such as CPU temperature it is necessary
to have an aggregation mode set. For some reason the default is None.
This change switches the three custom WPA 10 generic events views to
use Average as the aggregation mode for all the custom payload fields.

The other churn in the startup file is presumed to come from upgrading
to the RTM version of WPT 10.