Fixed glitches in the manifest compilation command
The manifest compilation command in the ETWProviders project had extra
spaces at the end of two of the commands which meant that nothing would
be shown when all-configurations/all-platforms were selected. Deleted.

The commands also used %(Filename)(%Extension) to specify the input file
and this breaks when the manifest file is not in the project directory.
It is better (and simpler!) to use %(FullPath) instead.

None of this makes a difference but it avoids future problems for people
who copy and modify the project.