Simplified recording of Chrome ETW events
In order to record ETW events from Chrome it is necessary to register
the Chrome ETW provider. The official way to do this is to run wevtutil
on chrome.dll. However this requires knowing where it is located which
is tricky given that we have stable, beta, dev, canary, and locally
built variations.

It turns out that you don't need to use the actual chrome.dll. You just
need a DLL that has Chrome's manifest compiled into it. This change adds
DummyChrome which is such a DLL, and it changes (simplifies!) UIforETW
to automatically use this DLL.

This gets rid of one of the prerequisites for recording ETW events
from Chrome. Another prerequisite was that Chrome developer had to be
checked. That was a bad idea so that has been removed.

The two remaining requirements are that you have to have some categories
selected in UIforETW's settings dialog (that makes sense) and you have
to add --trace-export-events-to-etw to Chrome's command-line. Selecting
categories is a logical necessity, but some day we plan to get rid of
the need for the command-line option.