Update WPA 10 startup profile for easier graphing
In order to make it easier to graph generic events data as discussed in
https://randomascii.wordpress.com/2015/08/03/graph-all-the-things-using-wpt-10/
this change adds custom presets for graphing fields 1-6 of generic
events data. Here's how to use:

Explore the generic events data, drilling down into an interesting area.
Rearrange columns as necessary so that you can group the data that you
need. For instance, if you want to graph processor power usage you would
drill down into UIforETW, Multi-Main, PowerStatus, and then drag Field 1
(MSR name in this context) to the left of the orange bar. Then right
click 'Processor' and select Filter to Selection.

Now that you've isolated the data you are interested in you need can
graph it. Since the Power (W) data is in Field 2 you would select the
"Randomascii Graph Field 2" preset. This moves Field 2 to the graphing
column and changes the graph type to stacked bars.

Simple!