Tweaked RetrieveSymbols
RetrieveSymbols would fail to build when TESTING was defined due to a
warning and to building with warnings-as-errors. I suppressed the
warning.

RetrieveSymbols would fail without explanation if _NT_SYMBOL_PATH was
not set - now it prints _NT_SYMBOL_PATH and warns if it isn't set.

RetrieveSymbols wasn't building to the bin directory - fixed.