Made Chrome developer default to off
When bChromeDeveloper_ is set then Chrome's symbol server path is added
to _NT_SYMBOL_PATH, StripChromeSymbols is called, and
IdentifyChromeProcesses.py is run. StripChromeSymbols can be
significantly expensive (many minutes sometimes, especially on a slow
Internet connection) and it is unnecessary for somebody who just wants
to record a trace.

Given the push to use UIforETW as a simple way for non-experts to
record traces, and given the use of UIforETW in many non-Chrome dev
shops, it does not make sense for bChromeDeveloper_ to default to true.