Added Identify Chrome Processes to the trace context menu
When Chrome Developer is not selected it can still be nice to be able to
get a list of Chrome processes, grouped by type. This change adds a
context menu option to do this.

It also deletes the C++ implementation of IdentifyChromeProcesses
because it is out of date and not needed.

Also fixed a bug in SaveNotesIfNeeded where it wasn't updating
traceNotes_ with the new saved content which led to accidental
saving of data when not wanted.