Proper support for launching python.bat or python.exe
Much to my surprise CreateProcess can launch batch files with no
difficulty, which makes supporting depot_tools' python.bat a trivial
exercise. This change mostly removes the misguided code from the
previous change, and adds a tiny bit of code to look for and use
python.bat if python.exe is not found.