Filter out 75% of the Microsoft-Windows-Win32k events.
Three critical section related events were causing 75% of the traffic
for the Microsoft-Windows-Win32k on Windows 7. This was causing the
buffers to wrap around extremely quickly, causing the user events
to cover a very short period of time, often not overlapping at all
with the kernel events.

This is probably also why the window-in-focus graph is frequently
empty.

The 0xFFFFFF was experimentally determined based on two tests. A more
refined filter may get released at some point.