Allow creation of flame graphs from UIforETW menu
Flame graphs can be handy for summarizing CPU usage but if they are too
difficult to use they won't get used. This is a first attempt at making
them more accessible. This adds a context menu that creates a flame
graph for the busiest thread.

Future work could include specifying a time range, process of interest,
thread of interest, number of threads to report on, or threshold - so
many choices! It would be better if this was burned in to WPA, but oh
well.

This also rearranges the context menu by putting scripts in a sub-menu
and compression in another - it was getting too big.