Add display of raw window focus events
The Window in Focus graph in WPA is very useful for finding key events
in a trace, but in WPA 10 it shows up quite unreliably - rarely, in
fact. So, this change adds a new View Preset to the Generic Events
table and shows the focus events by default. They aren't graphed as
nicely but the raw payload can be interpreted - open the table if the
graph information isn't sufficient.