Check default trace viewer each time
When UIforETW is run the first time it may take a few seconds for the
different versions of WPT to install. This leads to a race condition
where WPT 10 may be seen as not installed. By checking for its
existence when the viewer is needed the race condition is made less
severe.

The ideal fix would be to wait for the installs to finish but since
I can't uninstall I have no way to test that.