Removed usage of MFC "feature pack" classes.
The primary goal of this is to remove dependencies on DLLs that are not included in the list of DLLs that are supported on the Windows Server Core configurations. This removes the dependencies on MSIMG32.dll, gdiplus.dll, and WINMM.dll. It also gets rid of a dependency on imm32.dll, but that one is actually supported.

As a side effect, this also reduces the static exe size by about 2 megabytes, but the actual benefit of that ends up being minimal.