Add a tool to generate a flame graph from an ETW trace.
A flame graph is a tool to visualize frequent call stacks efficiently.
Contrary to the existing flame graph functionality of UIforETW, this
tool generates a flame graph that includes off-CPU time. Knowledge of
Chrome is bundled into the tool to ignore uninteresting off-CPU time
(e.g. message loop waiting for a new task to execute).

Usage of the tool is described in ETWInsights\README.md.

Code to read an ETW trace that is not specific to flame graph
generation resides in the ETWInsights\etw_reader, as it could be
reused to build other tools (e.g. an ETW to chrome://tracing
converter).