Call GetLastError as soon as possible for VC++ 2015 compat
The memory allocation functions used by VC++ 2015 debug builds zero the last
error field when you allocate memory. See this bug for details:

https://connect.microsoft.com/VisualStudio/feedback/details/1775690

The first time I encountered this it was a horror-show of debugging, but this
time (simpler scenario, trivial repro, familiar bug) it was a trivial fix.

In general it is a good idea to call GetLastError immediately after the function
that sets it, but I wish VC++ 2015 wouldn't force that behavior quite so
strongly.