Improve WPT installation from UIforETW
UIforETW has been installing Windows Performance Toolkit (WPT) version
10 (and version 8.1 as needed). However it did this using ShellExecute
which makes the installation asynchronous, which makes for a confusing
out-of-box experience.

Using msiexec.exe means that we automatically wait for the installation.
It also means that we can easily print the command executed, and its
output, all of which makes it clearer what is going on.