Disable spurious C6054 warning and remove test function.
IntentionalError() was added to check whether /analyze was working since
it was briefly broken in the run-up to Update 1. /analyze is now working
again. However VS 2015's header files annotate some of the _s functions
badly so that correct usage of them triggers C6054 warnings, so I am
forced to disable these warnings.