Upgrade UIforETW to Windows 10 TH2 WPT
The Windows 10 TH2 version of WPT fixes some serious problems including
not recording debug information on Windows 7 and crashing when modifying
the symbol load settings. Rather than support a complex matrix of many
different versions I am making it so that UIforETW requires the latest
version of WPT - 10.0.15086.15 - and has it as the default for
everything. WPT 8.1 is still supported, in case anyone wants to launch
WPA 8.1, but that is all it is used for now.