Fix various bugs in IdentifyChromeProcesses.py
IdentifyChromeProcesses.py was failing on some traces because exePath
did not necessary end with .exe, leading to no exePath being associated
with the browser process. This was fixed by looking for chrome.exe in
the processName field (parts[4]).

Then there was a problem with the crashpad-handler process type not
being identified because its exePath didn't end with a quote. The
regex was changed to not require that --type be the first parameter.

Also, the type matching is now done on commandLine rather than on the
entire line, because that makes more sense.

And, the child processes are now printed out indented, to be more
visually pleasing.