Add etw_cpuusage_longterm.bat
This batch file is designed for long-running recording of context switch
data and nothing else. This allows monitoring of all processes CPU usage
(including interrupts and DPCs) for periods of around an hour at a time.
The goal is to investigate badly-behaved processes (svchost.exe, I'm
looking at you) which spring to life at annoying times and waste battery
life.