Only update trace information text when needed
If you edit information for a trace, then alt+tab away, then when you
come back the cursor will have moved to the first character. This is
annoying when copying information from WPA to the trace information.

This was happening because when you alt+tab away UIforETW saves the
edited trace information - yay. Then the directory notification
thread notices that something has changed and it reloads the text.

This change just gets UIforETW to notice that the reload isn't really
needed so it can skip the call to SetDlgItemText.