Updates for better long-term ETW tracing
Use the "Circular Kernel Context Logger" instead of "NT Kernel Logger"
so that UIforETW can be used to record traces while long-term tracing is
running. Also, stop tracing on failure to start tracing, in order to
improve the odds of success.

Clean up the output to reduce noise, and print out a reasonable syntax
for viewing traces with a simple wpaProfile.