Tweaks to StripChromeSymbols.py
There seem to be some PDBs that UIforETW's copy of pdbcopy.exe can't
strip, so try using the Windows 10 SDK version if it is installed.

pdbcopy was effectively silently failing so now it fails more
visibly - the error output is displayed, and existence of the
destination file is checked.

Renaming of local PDBs would fail if the destination (.pdbx) file
already existed - delete the destination before renaming if it exists.

If renaming of PDBs failed then generating the symbol files is not
helpful - better to leave that for WPA or let the user fix the problem
and try stripping again.