Remove amcache.hve rename message and tweak metatrace text
The renaming of amcache.hve (a hack to work around slow trace merging
on some machines) always fails on most machines, and this doesn't seem
to matter, so printing this information is just confusing.

Also, tweak the text that suggests using metatrace.bat to make it less
confusing. Previously I updated just one of the three places that
mentions metatrace.bat.