Add checking for new versions
On startup UIforETW will now spawn a thread which will grab the releases
page and scrape the HTML looking for the latest version number. Crude,
but effective. The version checking can be disabled in the settings
dialog.

This means that Version.h should be updated everytime a new release is
made, since otherwise the version check will fire inappropriately.