Keep UIforETW idle when tracing is off
When UIforETW was not tracing it wouild leave all of its monitoring
threads running, even though the events created were just going to
the bit bucket. This change stops and starts the monintoring threads
as tracing stops and starts. It also only starts the over-long-trace
timer when tracing is running, and it changes its frequency from
every second to every thirty seconds.

This should have no effect on functionality, but makes it more
reasonable to leave UIforETW running at all times.

Note that if you leave circular buffer tracing running then all of the
monitoring threads will continue to run.

The one use-case that this change prevents is using UIforETW as an event
generator while recording those events using ETW batch files. I am not
aware of anyone ever using UIforETW in this way. If somebody wants this
functionality then it would be easy to add a button that would enable
events while tracing is off.

See this blog post for my rant about software that refuses to go idle:
https://randomascii.wordpress.com/2016/03/08/power-wastage-on-an-idle-laptop/