Hack IdentifyChromeProcesses.py to handle "missing" --type=
Due to crrev.com/1928863002 chrome started putting --type=gpu-process
about 3,600 characters into the command line - invisible to task manager
and to this scripts ETW parsing.

crrev.com/2013573002 fixes chrome so that it puts --type=gpu-process at
the beginning again, but the weird behavior may be around for a while,
so this hacks the script to handle this case.