Increment version number from 1.32 to 1.40
The code that checked for new versions was discarding the last digit of
the version numbers that it read from the releases web page. This means
that it wouldn't notice new versions until the 0.1 column changes. So
I'm changing it. Apologies for the hack. The version checking code has
been fixed so incrementing the 0.01 column will be sufficient next time.