Update startup profile
With WPT 10.0.14393.33 (Anniversary Edition) the Window-In-Focus graph
is finally fixed so it is no longer necessary to display the raw events
that a few experts could use to see which window is active. So the raw
events view is now disabled. Also, since GPU data is not recorded by
default I've hidden that by default. The default view is much less
cluttered.

You can copy over the new default profile by clicking "Copy startup
profiles" in the settings dialog after you run UIforETW.