Check for existence and version of symsrv.dll
The Windows 10 Anniversary Edition of WPA requires that symsrv.dll from
the Debuggers directory be installed. This will normally happen
automatically (when the Windows SDK or VS 2015 are installed) but it is
not actually guaranteed. This adds checks for the existence and a
minimum version of symsrv.dll in the Debuggers directory.

If a significant number of people hit this check then I could bundle
the installers, but for now I'm not going to because I don't want to
add another 26 MB to the release zip files.