Support Windows 7
Windows 7 can't run the Windows 10 Anniversary edition of WPT. So,
UIforETW needs to package to WPT installers and select them. It also
now warns if an unsupported version is installed. It also now checks to
see whether process creation failed, in order to make the Windows 7
failures (or future failures) more obvious.