Add Randomascii flame by Process, thread, Time Since Last
Most flame graphs are pretty easy to create and don't really need a
preset, but this one is very subtle. It requires going to View Editor,
Advanced, Graph Configuration, removing the Start Time and Duration
rows and then adding in Start Time/Last Switch-Out Time and
Duration/Time Since Last rows. Not obvious or easy. So now it's a
preset.