Get StripChromeSymbols.py to download all Chrome symbols
WPA 10.0.14393 has a bug (an extra slash in its symbol download URLs
that means it can't download symbols from Chrome's symbol server.
Therefore it becomes important for this script to download all of
Chrome's symbols, instead of just the ones that are slow to
convert to .symcache files.

Ironically, the .symcache translation process has been dramatically
improved in WPA 10.0.14393, and I was hoping to stop using it, but this
new bug makes it more important than ever.