Add EventEmitter project, for lab usage emittin of ETW events
UIforETW handles the recording and managing of traces, and it also emits
ETW events which are recorded by those traces. In a lab setting it is
useful to separate those functions. The EventEmitter.exe project reuses
the UIforETW event emitter classes to sit in the background emitting
ETW events, while responsibility for recording those events is delegated
to other batch files.

You *could* run this at the same time as UIforETW, but that would mostly
just mean that events would be emitted and recorded twice.

See ReadMe.txt for usage instructions and a to-do list.