Monitor Windows system-wide timer resolution events
UIforETW already monitors the system-wide Windows timer resolution, as
set by timeBeginPeriod and NtSetTimerResolution. However that is based
on sampling, so it misses any short-term changes to the timer frequency,
and it can't tell *who* has raised the timer frequency.

I fortuitously discovered that the Microsoft-Windows-Kernel-Power
provider - undocumented, like most of these - provides events whenever
the timer frequency changes. The timer changes are contained in events
such as SystemTimeResolutionChange and SystemTimeResolutionKernelChange
for changes by normal processes and the kernel respectively.

The new "Randomascii System Time Resolution" preset for the Generic
Events is set up with a filter to show just these events. To configure
this filtering I used the View Editor, Advanced, and the syntax that is
lightly documented here:
  https://msdn.microsoft.com/en-us/windows/hardware/commercialize/test/wpt/wpa-query-syntax
Look at SystemTimeResolutionChange and SystemTimeResolutionKernelChange, and in
 particular at the RequestedResolution field (units are 0.1 microseconds, so
0x2710 == 10,000 = 1 ms).