Make trace starting more robust
Starting of ETW tracing will fail if there is another trace running with
the same kernel logger (and there are only two), with the same name for
a user logger sesion, or if the file names being used for tracing to a
file are in use. This means that starting tracing will fail if it wasn't
stopped for some reason (UIforETW crash perhaps) and tracing may fail
if some other program happens to be using the same file or session
names. These start failures are not self healing - repeated attempts to
start tracing will repeatedly fail. Previously UIforETW dealt with this
by printing a message suggesting that you stop tracing. This didn't
always work.

This change makes trace starting far more robust by unilaterally
stopping all trace sessions that UIforETW might use. This change also
changes the session names and file names to make them less likely to
collide.

If this doesn't work then changing the kernel logger using the checkbox
added in the previous change may help.