Add no-wrap attribute to static text to avoid wrapping
Issue #96 points out that at some scaling levels (such as 125%) the text
in some static text controls in UIforETW partially disappears. This is
because the static text controls are very tightly fit, and because
text wrappin is enabled. I also widened some of the text controls to
make clipping less likely.

<rant>The relevant flag is displayed in the properties dialog in Visual
Studio as "No Wrap" with a default of False, which means that wrapping
is True, and it means that Visual Studio loves double negatives, despite
whatever confusion they might cause. And, the default is probably
incorrect given that most static text controls are single line.</rant>