Record VirtualFree stacks when VirtualAlloc stacks are requested
VirtualFree stacks can be used to track down memory leaks if the memory
is freed during shutdown (not true leaks, but you know what I mean).
This change adds VirtualFree to the stack walk flags when VirtualAlloc
is requested. This change also fixes the capitalization of the other
stack walk flags, to match that used by xperf -help stackwalk