Allow using Ctrl+Win+C to start and stop tracing
Ctrl+Win+C has long been the global hot key which UIforETW sets up for
recording traces. This allows recording a trace to disk without leaving
the application that you are profiling. Sometimes it is also necessary
to *start* tracing without leaving the application of interest. This
change allows that.