Stripped down CPUUsageByProcess.wpaProfile
When creating .wpaProfile files they naturally contain lots of
excess junk, such as every other custom view in your startup profile,
modified profiles, etc. The best way to avoid this seems to be
manual editing of the .wpaProfile files, or deleting the startup
profile and other settings files before creating profiles.

Failure to do this isn't fatal, but having ~180 KB of pointless
XML seems wasteful, and obscures the purpose of the file.

This change strips CPUUsageByProcess.wpaProfile from 219,934 bytes to
26,877 bytes.